django-app/requirements.txt
Django==4.2.13
django-cors-headers==4.3.1
django-split-settings==1.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
python-dotenv==1.0.1
PyYAML==5.3.1
gunicorn